pyproject.toml
[project]
name = "billy"
version = "0.1.0"
description = ""
authors = [
    {name = "Erwin",email = "erwinwingyonata@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "numpy (>=2.2.6,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "requests (>=2.32.4,<3.0.0)",
    "redis[hiredis] (>=6.2.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
